SDI-REST_API-main/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
idna==3.4
pydot==1.4.2
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
